chore: remove gov proposals from proposal_handler.go
