+++ app/proposal_handler.go
@@ -3,11 +3,6 @@ package app
 import (
 	"cosmossdk.io/log"
 	bundlestypes "github.com/KYVENetwork/chain/x/bundles/types"
-	delegationtypes "github.com/KYVENetwork/chain/x/delegation/types"
-	funderstypes "github.com/KYVENetwork/chain/x/funders/types"
-	globaltypes "github.com/KYVENetwork/chain/x/global/types"
-	pooltypes "github.com/KYVENetwork/chain/x/pool/types"
-	stakerstypes "github.com/KYVENetwork/chain/x/stakers/types"
 	abci "github.com/cometbft/cometbft/abci/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"reflect"
@@ -26,21 +21,6 @@ func NewPriorityProposalHandler(logger log.Logger, decoder sdk.TxDecoder) *Prior
 	}
 }
 
-var govTypes = []string{
-	reflect.TypeOf(bundlestypes.MsgUpdateParams{}).Name(),
-	reflect.TypeOf(delegationtypes.MsgUpdateParams{}).Name(),
-	reflect.TypeOf(funderstypes.MsgUpdateParams{}).Name(),
-	reflect.TypeOf(globaltypes.MsgUpdateParams{}).Name(),
-	reflect.TypeOf(pooltypes.MsgCreatePool{}).Name(),
-	reflect.TypeOf(pooltypes.MsgUpdatePool{}).Name(),
-	reflect.TypeOf(pooltypes.MsgDisablePool{}).Name(),
-	reflect.TypeOf(pooltypes.MsgEnablePool{}).Name(),
-	reflect.TypeOf(pooltypes.MsgScheduleRuntimeUpgrade{}).Name(),
-	reflect.TypeOf(pooltypes.MsgCancelRuntimeUpgrade{}).Name(),
-	reflect.TypeOf(pooltypes.MsgUpdateParams{}).Name(),
-	reflect.TypeOf(stakerstypes.MsgUpdateParams{}).Name(),
-}
-
 var priorityTypes = []string{
 	reflect.TypeOf(bundlestypes.MsgSubmitBundleProposal{}).Name(),
 	reflect.TypeOf(bundlestypes.MsgVoteBundleProposal{}).Name(),
@@ -53,12 +33,10 @@ var priorityTypes = []string{
 func (h *PriorityProposalHandler) PrepareProposal() sdk.PrepareProposalHandler {
 	return func(ctx sdk.Context, req *abci.RequestPrepareProposal) (*abci.ResponsePrepareProposal, error) {
 		// Separate the transactions into different queues
-		// govQueue: transactions that can only be executed by the governance module
 		// priorityQueue: transactions that should be executed before the default transactions
 		// defaultQueue: transactions that should be executed last
-		// The order is govQueue -> priorityQueue -> defaultQueue
+		// The order is priorityQueue -> defaultQueue
 
-		var govQueue [][]byte
 		var priorityQueue [][]byte
 		var defaultQueue [][]byte
 
@@ -80,27 +58,21 @@ func (h *PriorityProposalHandler) PrepareProposal() sdk.PrepareProposalHandler {
 				msg := msgs[0]
 				msgType := string(msg.ProtoReflect().Type().Descriptor().Name())
 
-				if slices.Contains(govTypes, msgType) {
-					govQueue = append(govQueue, rawTx)
-					continue
-				}
-
 				if slices.Contains(priorityTypes, msgType) {
 					priorityQueue = append(priorityQueue, rawTx)
 					continue
 				}
 			}
 
-			// Otherwise, add the message to the default queue
+			// Otherwise, add the tx to the default queue
 			defaultQueue = append(defaultQueue, rawTx)
 		}
 
-		// Combine all the queues
-		newTxs := append(govQueue, priorityQueue...)
-		newTxs = append(newTxs, defaultQueue...)
+		// Append the default queue to the priority queue
+		priorityQueue = append(priorityQueue, defaultQueue...)
 
 		return &abci.ResponsePrepareProposal{
-			Txs: newTxs,
+			Txs: priorityQueue,
 		}, nil
 	}
 }
